improve delete method with 404 error when user doesn't exist [S08_X]

+++ src/restful-web-services-06/restful-web-services/src/main/java/com/in28minutes/rest/webservices/restfulwebservices/user/UserController.java
@@ -52,6 +52,10 @@ public class UserController {
 	
 	@DeleteMapping("/users/{id}")
 	public void deleteUser(@PathVariable int id) {
+		Optional<User> user = service.findById(id);
+		if (user.isEmpty()) {
+			throw new UserNotFoundException("id:"+id);
+		}
 		service.deleteById(id);
 	}
